refactor(core): move machine tests out of manager
